@@ -2,13 +2,16 @@ import { browserHistory } from 'react-router';
 import axios from 'axios';
 import objectID from 'bson-objectid';
 import each from 'async/each';
-import { isEqual } from 'lodash';
+import isEqual from 'lodash/isEqual';
 import * as ActionTypes from '../../../constants';
 import { showToast, setToastText } from './toast';
-import { setUnsavedChanges,
+import {
+  setUnsavedChanges,
   justOpenedProject,
   resetJustOpenedProject,
-  showErrorModal } from './ide';
+  showErrorModal,
+  setPreviousPath
+} from './ide';
 import { clearState, saveState } from '../../../persistState';
 
 const __process = (typeof global !== 'undefined' ? global : window).process;
@@ -27,6 +30,22 @@ export function setProjectName(name) {
   return {
     type: ActionTypes.SET_PROJECT_NAME,
     name
+  };
+}
+
+export function projectSaveFail(error) {
+  return {
+    type: ActionTypes.PROJECT_SAVE_FAIL,
+    error
+  };
+}
+
+export function setNewProject(project) {
+  return {
+    type: ActionTypes.NEW_PROJECT,
+    project,
+    owner: project.user,
+    files: project.files
   };
 }
 
@@ -64,29 +83,71 @@ export function clearPersistedState() {
   };
 }
 
-export function saveProject(autosave = false) {
+export function startSavingProject() {
+  return {
+    type: ActionTypes.START_SAVING_PROJECT
+  };
+}
+
+export function endSavingProject() {
+  return {
+    type: ActionTypes.END_SAVING_PROJECT
+  };
+}
+
+export function projectSaveSuccess() {
+  return {
+    type: ActionTypes.PROJECT_SAVE_SUCCESS
+  };
+}
+
+// want a function that will check for changes on the front end
+function getSynchedProject(currentState, responseProject) {
+  let hasChanges = false;
+  const synchedProject = Object.assign({}, responseProject);
+  const currentFiles = currentState.files.map(({ name, children, content }) => ({ name, children, content }));
+  const responseFiles = responseProject.files.map(({ name, children, content }) => ({ name, children, content }));
+  if (!isEqual(currentFiles, responseFiles)) {
+    synchedProject.files = currentState.files;
+    hasChanges = true;
+  }
+  if (currentState.project.name !== responseProject.name) {
+    synchedProject.name = currentState.project.name;
+    hasChanges = true;
+  }
+  return {
+    synchedProject,
+    hasChanges
+  };
+}
+
+export function saveProject(selectedFile = null, autosave = false) {
   return (dispatch, getState) => {
     const state = getState();
+    if (state.project.isSaving) {
+      return Promise.resolve();
+    }
+    dispatch(startSavingProject());
     if (state.user.id && state.project.owner && state.project.owner.id !== state.user.id) {
       return Promise.reject();
     }
     const formParams = Object.assign({}, state.project);
     formParams.files = [...state.files];
+    if (selectedFile) {
+      const fileToUpdate = formParams.files.find(file => file.id === selectedFile.id);
+      fileToUpdate.content = selectedFile.content;
+    }
     if (state.project.id) {
       return axios.put(`${ROOT_URL}/projects/${state.project.id}`, formParams, { withCredentials: true })
         .then((response) => {
-          const currentState = getState();
-          const savedProject = Object.assign({}, response.data);
-          if (!isEqual(currentState.files, response.data.files)) {
-            savedProject.files = currentState.files;
+          dispatch(endSavingProject());
+          dispatch(setUnsavedChanges(false));
+          const { hasChanges, synchedProject } = getSynchedProject(getState(), response.data);
+          if (hasChanges) {
             dispatch(setUnsavedChanges(true));
-          } else {
-            dispatch(setUnsavedChanges(false));
           }
-          dispatch(setProject(savedProject));
-          dispatch({
-            type: ActionTypes.PROJECT_SAVE_SUCCESS
-          });
+          dispatch(setProject(synchedProject));
+          dispatch(projectSaveSuccess());
           if (!autosave) {
             if (state.ide.justOpenedProject && state.preferences.autosave) {
               dispatch(showToast(5500));
@@ -100,30 +161,32 @@ export function saveProject(autosave = false) {
           }
         })
         .catch((response) => {
+          dispatch(endSavingProject());
           if (response.status === 403) {
             dispatch(showErrorModal('staleSession'));
           } else if (response.status === 409) {
             dispatch(showErrorModal('staleProject'));
           } else {
-            dispatch({
-              type: ActionTypes.PROJECT_SAVE_FAIL,
-              error: response.data
-            });
+            dispatch(projectSaveFail(response.data));
           }
         });
     }
 
     return axios.post(`${ROOT_URL}/projects`, formParams, { withCredentials: true })
       .then((response) => {
-        dispatch(setUnsavedChanges(false));
-        dispatch(setProject(response.data));
-        browserHistory.push(`/${response.data.user.username}/sketches/${response.data.id}`);
-        dispatch({
-          type: ActionTypes.NEW_PROJECT,
-          project: response.data,
-          owner: response.data.user,
-          files: response.data.files
-        });
+        dispatch(endSavingProject());
+        const { hasChanges, synchedProject } = getSynchedProject(getState(), response.data);
+        if (hasChanges) {
+          dispatch(setNewProject(synchedProject));
+          dispatch(setUnsavedChanges(false));
+          browserHistory.push(`/${response.data.user.username}/sketches/${response.data.id}`);
+          dispatch(setUnsavedChanges(true));
+        } else {
+          dispatch(setNewProject(synchedProject));
+          dispatch(setUnsavedChanges(false));
+          browserHistory.push(`/${response.data.user.username}/sketches/${response.data.id}`);
+        }
+        dispatch(projectSaveSuccess());
         if (!autosave) {
           if (state.preferences.autosave) {
             dispatch(showToast(5500));
@@ -137,13 +200,11 @@ export function saveProject(autosave = false) {
         }
       })
       .catch((response) => {
+        dispatch(endSavingProject());
         if (response.status === 403) {
           dispatch(showErrorModal('staleSession'));
         } else {
-          dispatch({
-            type: ActionTypes.PROJECT_SAVE_FAIL,
-            error: response.data
-          });
+          dispatch(projectSaveFail(response.data));
         }
       });
   };
@@ -151,27 +212,7 @@ export function saveProject(autosave = false) {
 
 export function autosaveProject() {
   return (dispatch, getState) => {
-    saveProject(true)(dispatch, getState);
-  };
-}
-
-export function createProject() {
-  return (dispatch) => {
-    axios.post(`${ROOT_URL}/projects`, {}, { withCredentials: true })
-      .then((response) => {
-        browserHistory.push(`/${response.data.user.username}/sketches/${response.data.id}`);
-        dispatch({
-          type: ActionTypes.NEW_PROJECT,
-          project: response.data,
-          owner: response.data.user,
-          files: response.data.files
-        });
-        dispatch(setUnsavedChanges(false));
-      })
-      .catch(response => dispatch({
-        type: ActionTypes.PROJECT_SAVE_FAIL,
-        error: response.data
-      }));
+    saveProject(null, true)(dispatch, getState);
   };
 }
 
@@ -206,52 +247,61 @@ function generateNewIdsForChildren(file, files) {
   file.children = newChildren; // eslint-disable-line
 }
 
-export function cloneProject() {
+export function cloneProject(id) {
   return (dispatch, getState) => {
     dispatch(setUnsavedChanges(false));
-    const state = getState();
-    const newFiles = state.files.map((file) => { // eslint-disable-line
-      return { ...file };
-    });
-
-    // generate new IDS for all files
-    const rootFile = newFiles.find(file => file.name === 'root');
-    const newRootFileId = objectID().toHexString();
-    rootFile.id = newRootFileId;
-    rootFile._id = newRootFileId;
-    generateNewIdsForChildren(rootFile, newFiles);
-
-    // duplicate all files hosted on S3
-    each(newFiles, (file, callback) => {
-      if (file.url && file.url.includes('amazonaws')) {
-        const formParams = {
-          url: file.url
-        };
-        axios.post(`${ROOT_URL}/S3/copy`, formParams, { withCredentials: true })
-          .then((response) => {
-            file.url = response.data.url;
-            callback(null);
-          });
+    new Promise((resolve, reject) => {
+      if (!id) {
+        resolve(getState());
       } else {
-        callback(null);
+        fetch(`${ROOT_URL}/projects/${id}`)
+          .then(res => res.json())
+          .then(data => resolve({
+            files: data.files,
+            project: {
+              name: data.name
+            }
+          }));
       }
-    }, (err) => {
-      // if not errors in duplicating the files on S3, then duplicate it
-      const formParams = Object.assign({}, { name: `${state.project.name} copy` }, { files: newFiles });
-      axios.post(`${ROOT_URL}/projects`, formParams, { withCredentials: true })
-        .then((response) => {
-          browserHistory.push(`/${response.data.user.username}/sketches/${response.data.id}`);
-          dispatch({
-            type: ActionTypes.NEW_PROJECT,
-            project: response.data,
-            owner: response.data.user,
-            files: response.data.files
-          });
-        })
-        .catch(response => dispatch({
-          type: ActionTypes.PROJECT_SAVE_FAIL,
-          error: response.data
-        }));
+    }).then((state) => {
+      const newFiles = state.files.map((file) => { // eslint-disable-line
+        return { ...file };
+      });
+
+      // generate new IDS for all files
+      const rootFile = newFiles.find(file => file.name === 'root');
+      const newRootFileId = objectID().toHexString();
+      rootFile.id = newRootFileId;
+      rootFile._id = newRootFileId;
+      generateNewIdsForChildren(rootFile, newFiles);
+
+      // duplicate all files hosted on S3
+      each(newFiles, (file, callback) => {
+        if (file.url && file.url.includes('amazonaws')) {
+          const formParams = {
+            url: file.url
+          };
+          axios.post(`${ROOT_URL}/S3/copy`, formParams, { withCredentials: true })
+            .then((response) => {
+              file.url = response.data.url;
+              callback(null);
+            });
+        } else {
+          callback(null);
+        }
+      }, (err) => {
+        // if not errors in duplicating the files on S3, then duplicate it
+        const formParams = Object.assign({}, { name: `${state.project.name} copy` }, { files: newFiles });
+        axios.post(`${ROOT_URL}/projects`, formParams, { withCredentials: true })
+          .then((response) => {
+            browserHistory.push(`/${response.data.user.username}/sketches/${response.data.id}`);
+            dispatch(setNewProject(response.data));
+          })
+          .catch(response => dispatch({
+            type: ActionTypes.PROJECT_SAVE_FAIL,
+            error: response.data
+          }));
+      });
     });
   };
 }
@@ -265,5 +315,67 @@ export function showEditProjectName() {
 export function hideEditProjectName() {
   return {
     type: ActionTypes.HIDE_EDIT_PROJECT_NAME
+  };
+}
+
+export function setProjectSavedTime(updatedAt) {
+  return {
+    type: ActionTypes.SET_PROJECT_SAVED_TIME,
+    value: updatedAt
+  };
+}
+
+export function changeProjectName(id, newName) {
+  return (dispatch, getState) => {
+    const state = getState();
+    axios.put(`${ROOT_URL}/projects/${id}`, { name: newName }, { withCredentials: true })
+      .then((response) => {
+        if (response.status === 200) {
+          dispatch({
+            type: ActionTypes.RENAME_PROJECT,
+            payload: { id: response.data.id, name: response.data.name }
+          });
+          if (state.project.id === response.data.id) {
+            dispatch({
+              type: ActionTypes.SET_PROJECT_NAME,
+              name: response.data.name
+            });
+          }
+        }
+      })
+      .catch((response) => {
+        console.log(response);
+        dispatch({
+          type: ActionTypes.PROJECT_SAVE_FAIL,
+          error: response.data
+        });
+      });
+  };
+}
+
+export function deleteProject(id) {
+  return (dispatch, getState) => {
+    axios.delete(`${ROOT_URL}/projects/${id}`, { withCredentials: true })
+      .then(() => {
+        const state = getState();
+        if (id === state.project.id) {
+          dispatch(resetProject());
+          dispatch(setPreviousPath('/'));
+        }
+        dispatch({
+          type: ActionTypes.DELETE_PROJECT,
+          id
+        });
+      })
+      .catch((response) => {
+        if (response.status === 403) {
+          dispatch(showErrorModal('staleSession'));
+        } else {
+          dispatch({
+            type: ActionTypes.ERROR,
+            error: response.data
+          });
+        }
+      });
   };
 }
