@@ -5,6 +5,7 @@ import classNames from 'classnames';
 import InlineSVG from 'react-inlinesvg';
 import NewFileForm from './NewFileForm';
 import FileUploader from './FileUploader';
+import { CREATE_FILE_REGEX } from '../../../../server/utils/fileUtils';
 
 const exitUrl = require('../../../images/exit.svg');
 
@@ -71,8 +72,8 @@ function validate(formProps) {
 
   if (!formProps.name) {
     errors.name = 'Please enter a name';
-  } else if (!formProps.name.match(/(.+\.js$|.+\.css$|.+\.json$|.+\.txt$|.+\.csv$|.+\.tsv$)/i)) {
-    errors.name = 'File must be of type JavaScript, CSS, JSON, TXT, CSV, or TSV.';
+  } else if (!formProps.name.match(CREATE_FILE_REGEX)) {
+    errors.name = 'Invalid file type. Valid extensions are .js, .css, .json, .txt, .csv, .tsv, .frag, and .vert.';
   }
 
   return errors;
