@@ -169,10 +169,20 @@ class PreviewFrame extends React.Component {
     return newContent;
   }
 
+  mergeLocalFilesAndEditorActiveFile() {
+    const files = this.props.files.slice();
+    if (this.props.cmController.getContent) {
+      const activeFileInEditor = this.props.cmController.getContent();
+      files.find(file => file.id === activeFileInEditor.id).content = activeFileInEditor.content;
+    }
+    return files;
+  }
+
   injectLocalFiles() {
     const htmlFile = this.props.htmlFile.content;
     let scriptOffs = [];
-    const resolvedFiles = this.resolveJSAndCSSLinks(this.props.files);
+    const files = this.mergeLocalFilesAndEditorActiveFile();
+    const resolvedFiles = this.resolveJSAndCSSLinks(files);
     const parser = new DOMParser();
     const sketchDoc = parser.parseFromString(htmlFile, 'text/html');
 
@@ -350,6 +360,7 @@ class PreviewFrame extends React.Component {
   }
 
   renderSketch() {
+    this.props.clearConsole();
     const doc = this.iframeElement;
     const localFiles = this.injectLocalFiles();
     if (this.props.isPlaying) {
@@ -404,11 +415,16 @@ PreviewFrame.propTypes = {
   fullView: PropTypes.bool,
   setBlobUrl: PropTypes.func.isRequired,
   stopSketch: PropTypes.func.isRequired,
-  expandConsole: PropTypes.func.isRequired
+  expandConsole: PropTypes.func.isRequired,
+  clearConsole: PropTypes.func.isRequired,
+  cmController: PropTypes.shape({
+    getContent: PropTypes.func
+  })
 };
 
 PreviewFrame.defaultProps = {
-  fullView: false
+  fullView: false,
+  cmController: {}
 };
 
 export default PreviewFrame;
