@@ -1,9 +1,15 @@
 import * as ActionTypes from '../../../constants';
 
-const assets = (state = [], action) => {
+// 1,000,000 bytes in a MB. can't upload if totalSize is bigger than this.
+const initialState = {
+  list: [],
+  totalSize: 0
+};
+
+const assets = (state = initialState, action) => {
   switch (action.type) {
     case ActionTypes.SET_ASSETS:
-      return action.assets;
+      return { list: action.assets, totalSize: action.totalSize };
     default:
       return state;
   }
