@@ -1,12 +1,19 @@
-import generate from 'project-name-generator';
+import friendlyWords from 'friendly-words';
 import * as ActionTypes from '../../../constants';
 
+const generateRandomName = () => {
+  const adj = friendlyWords.predicates[Math.floor(Math.random() * friendlyWords.predicates.length)];
+  const obj = friendlyWords.objects[Math.floor(Math.random() * friendlyWords.objects.length)];
+  return `${adj} ${obj}`;
+};
+
 const initialState = () => {
-  const generatedString = generate({ words: 2 }).spaced;
+  const generatedString = generateRandomName();
   const generatedName = generatedString.charAt(0).toUpperCase() + generatedString.slice(1);
   return {
     name: generatedName,
-    updatedAt: ''
+    updatedAt: '',
+    isSaving: false
   };
 };
 
@@ -22,14 +29,16 @@ const project = (state, action) => {
         id: action.project.id,
         name: action.project.name,
         updatedAt: action.project.updatedAt,
-        owner: action.owner
+        owner: action.owner,
+        isSaving: false
       };
     case ActionTypes.SET_PROJECT:
       return {
         id: action.project.id,
         name: action.project.name,
         updatedAt: action.project.updatedAt,
-        owner: action.owner
+        owner: action.owner,
+        isSaving: false
       };
     case ActionTypes.RESET_PROJECT:
       return initialState();
@@ -39,6 +48,10 @@ const project = (state, action) => {
       return Object.assign({}, state, { isEditingName: false });
     case ActionTypes.SET_PROJECT_SAVED_TIME:
       return Object.assign({}, state, { updatedAt: action.value });
+    case ActionTypes.START_SAVING_PROJECT:
+      return Object.assign({}, state, { isSaving: true });
+    case ActionTypes.END_SAVING_PROJECT:
+      return Object.assign({}, state, { isSaving: false });
     default:
       return state;
   }
