@@ -10,6 +10,9 @@ import sketches from './modules/IDE/reducers/projects';
 import toast from './modules/IDE/reducers/toast';
 import console from './modules/IDE/reducers/console';
 import assets from './modules/IDE/reducers/assets';
+import search from './modules/IDE/reducers/search';
+import sorting from './modules/IDE/reducers/sorting';
+import loading from './modules/IDE/reducers/loading';
 
 const rootReducer = combineReducers({
   form,
@@ -19,10 +22,13 @@ const rootReducer = combineReducers({
   user,
   project,
   sketches,
+  search,
+  sorting,
   editorAccessibility,
   toast,
   console,
-  assets
+  assets,
+  loading
 });
 
 export default rootReducer;
