@@ -15,6 +15,9 @@ export const domOnlyProps = ({
   ...domProps }) => domProps;
 /* eslint-enable */
 
+// eslint-disable-next-line max-len
+const EMAIL_REGEX = /^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/i;
+
 function validateNameEmail(formProps, errors) {
   if (!formProps.username) {
     errors.username = 'Please enter a username.';
@@ -28,7 +31,7 @@ function validateNameEmail(formProps, errors) {
     errors.email = 'Please enter an email.';
   } else if (
     // eslint-disable-next-line max-len
-    !formProps.email.match(/^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/i)) {
+    !formProps.email.match(EMAIL_REGEX)) {
     errors.email = 'Please enter a valid email address.';
   }
 }
@@ -64,13 +67,29 @@ export function validateSignup(formProps) {
   if (!formProps.password) {
     errors.password = 'Please enter a password';
   }
+  if (formProps.password) {
+    if (formProps.password.length < 6) {
+      errors.password = 'Password must be at least 6 characters';
+    }
+  }
   if (!formProps.confirmPassword) {
     errors.confirmPassword = 'Please enter a password confirmation';
   }
 
-  if (formProps.password !== formProps.confirmPassword) {
-    errors.password = 'Passwords must match';
+  if (formProps.password !== formProps.confirmPassword && formProps.confirmPassword) {
+    errors.confirmPassword = 'Passwords must match';
   }
 
+  return errors;
+}
+export function validateResetPassword(formProps) {
+  const errors = {};
+  if (!formProps.email) {
+    errors.email = 'Please enter an email.';
+  } else if (
+    // eslint-disable-next-line max-len
+    !formProps.email.match(EMAIL_REGEX)) {
+    errors.email = 'Please enter a valid email address.';
+  }
   return errors;
 }
