@@ -1,3 +1,3 @@
-require('babel-register');
-require('babel-polyfill');
+require('@babel/register');
+require('@babel/polyfill');
 require('./db_reformat');
