@@ -1,5 +1,5 @@
-require('babel-register');
-require('babel-polyfill');
+require('@babel/register');
+require('@babel/polyfill');
 const dotenv = require('dotenv');
 
 if (process.env.NODE_ENV === 'development') {
