@@ -1,8 +1,9 @@
 const webpack = require('webpack');
 const path = require('path');
-const ExtractTextPlugin = require('extract-text-webpack-plugin');
+const TerserJSPlugin = require('terser-webpack-plugin');
+const MiniCssExtractPlugin = require('mini-css-extract-plugin');
+const OptimizeCSSAssetsPlugin = require('optimize-css-assets-webpack-plugin');
 const ManifestPlugin = require('webpack-manifest-plugin');
-const ChunkManifestPlugin = require('chunk-manifest-webpack-plugin');
 const cssnext = require('postcss-cssnext');
 const postcssFocus = require('postcss-focus');
 const postcssReporter = require('postcss-reporter');
@@ -13,35 +14,18 @@ if (process.env.NODE_ENV === "development") {
 
 module.exports = [{
   devtool: 'source-map',
-
+  mode: 'production',
   entry: {
     app: [
-      'babel-polyfill',
+      '@babel/polyfill',
+      'core-js/modules/es6.promise',
+      'core-js/modules/es6.array.iterator',
       path.resolve(__dirname, '../client/index.jsx')
-    ],
-    vendor: [
-      'axios',
-      'classnames',
-      'codemirror',
-      'csslint',
-      'dropzone',
-      'htmlhint',
-      'js-beautify',
-      'jshint',
-      'moment',
-      'react-dom',
-      'react-inlinesvg',
-      'react-redux',
-      'react-router',
-      'react',
-      'redux-form',
-      'redux-thunk',
-      'redux',
     ]
   },
   output: {
     path: path.resolve(__dirname, '../dist/static'),
-    filename: '[name].[chunkhash].js',
+    filename: '[name].[hash].js',
     publicPath: '/'
   },
 
@@ -52,16 +36,44 @@ module.exports = [{
       'node_modules',
     ]
   },
-
   module: {
-    rules: [
-      {
+    rules: [{
         test: /main\.scss$/,
         exclude: /node_modules/,
-        loader: ExtractTextPlugin.extract({
-          fallback: 'style-loader',
-          use: 'css-loader!sass-loader!postcss-loader'
-        })
+        use: [
+          MiniCssExtractPlugin.loader,
+          {
+            loader: 'css-loader',
+            options: {
+              importLoaders: 2,
+              sourceMap: true
+            }
+          },
+          {
+            loader: 'postcss-loader',
+            options: {
+              plugins: () => [
+                postcssFocus(),
+                cssnext({
+                  browsers: ['last 2 versions', 'IE > 9']
+                }),
+                cssnano({
+                  autoprefixer: false
+                }),
+                postcssReporter({
+                  clearMessages: true
+                })
+              ],
+              sourceMap: true
+            }
+          },
+          {
+            loader: 'sass-loader',
+            options: {
+              sourceMap: true
+            }
+          }
+        ]
       },
       {
         test: /\.jsx?$/,
@@ -75,11 +87,11 @@ module.exports = [{
       {
         test: /\.(png)$/,
         use: {
-            loader: 'file-loader',
-            options: {
-                name: '[name].[ext]',
-                outputPath: 'images/'
-            }
+          loader: 'file-loader',
+          options: {
+            name: '[name].[ext]',
+            outputPath: 'images/'
+          }
         }
       },
       {
@@ -91,56 +103,28 @@ module.exports = [{
         use: {
           loader: 'sass-extract-loader',
           options: {
-            plugins: [{ plugin: 'sass-extract-js', options: { camelCase: false } }]
+            plugins: [{
+              plugin: 'sass-extract-js',
+              options: {
+                camelCase: false
+              }
+            }]
           }
         }
       }
     ]
   },
-
+  optimization: {
+    minimizer: [new TerserJSPlugin({}), new OptimizeCSSAssetsPlugin({})],
+  },
   plugins: [
-    new webpack.DefinePlugin({
-      'process.env': {
-        NODE_ENV: JSON.stringify('production')
-      }
-    }),
-    new webpack.optimize.CommonsChunkPlugin({
-      name: 'vendor',
-      minChunks: Infinity,
-      filename: '[name].[chunkhash].js',
-    }),
-    new ExtractTextPlugin({ filename: 'app.[chunkhash].css', allChunks: true }),
     new ManifestPlugin({
       basePath: '/',
     }),
-    new ChunkManifestPlugin({
-      filename: 'chunk-manifest.json',
-      manifestVariable: 'webpackManifest',
-    }),
-    new webpack.optimize.UglifyJsPlugin({
-      sourceMap: true,
-      compress: {
-        warnings: false
-      }
-    }),
-    new webpack.LoaderOptionsPlugin({
-      options: {
-        postcss: () => [
-          postcssFocus(),
-          cssnext({
-            browsers: ['last 2 versions', 'IE > 9']
-          }),
-          cssnano({
-            autoprefixer: false
-          }),
-          postcssReporter({
-            clearMessages: true
-          })
-        ]
-      }
+    new MiniCssExtractPlugin({
+      filename: 'app.[hash].css',
     })
-  ],
-
+  ]
 },
 {
   entry: {
@@ -149,6 +133,7 @@ module.exports = [{
     ]
   },
   target: 'web',
+  mode: 'production',
   output: {
     path: path.resolve(__dirname, '../dist/static'),
     filename: 'previewScripts.js',
@@ -162,19 +147,15 @@ module.exports = [{
     ],
   },
   module: {
-    loaders: [
+    rules: [
       {
         test: /\.jsx?$/,
         exclude: /node_modules/,
-        loader: 'babel-loader'
+        loader: 'babel-loader',
+        options: {
+          babelrc: true
+        }
       }
     ]
-  },
-  plugins: [
-    new webpack.optimize.UglifyJsPlugin({
-      compress: {
-        warnings: false
-      }
-    })
-  ]
+  }
 }];
